fix: 자유게시판 목록 페이지 수정
 - 마지막 페이지 이동시 게시글 목록의 번호가 잘못 표기되었다가 다시 표기되는 현상이 있었습니다.
 - 페이지 이동 버튼 함수들의 코드 내 refetch() 함수 호출순서를 변경시키고, async/await 사용하여 해결했습니다.
