+++ src/commons/store/index.ts
@@ -6,11 +6,6 @@ export const accessTokenState = atom({
   default: "",
 });
 
-export const currentPageState = atom({
-  key: `currentPageState/${uuidv4()}`,
-  default: 1,
-});
-
 export const modalMessageState = atom({
   key: `modalMessageState/${uuidv4()}`,
   default: "",
