+++ src/components/units/board/list/BoardList.types.ts
@@ -4,7 +4,7 @@ import {
   type IQueryFetchBoardsCountArgs,
   type IQuery,
 } from "../../../../commons/types/generated/types";
-import { type MouseEvent, type Dispatch, type SetStateAction } from "react";
+import { type Dispatch, type SetStateAction } from "react";
 
 export interface IBoardListProps {
   data: Pick<IQuery, "fetchBoards" | "fetchBoardsCount"> | undefined;
@@ -18,6 +18,7 @@ export interface IBoardListProps {
   endPage: number;
   pages: number;
   keyword: string;
+  setCurrentPage: Dispatch<SetStateAction<number>>;
   setStartPage: Dispatch<SetStateAction<number>>;
   setKeyword: Dispatch<SetStateAction<string>>;
   onClickMoveToBoardDetail: (id: string) => () => void;
@@ -26,7 +27,7 @@ export interface IBoardListProps {
   onClickLastPage: () => void;
   onClickPrev: () => void;
   onClickNext: () => void;
-  onClickButton: (event: MouseEvent<HTMLDivElement>) => void;
+  onClickButton: (page: number) => () => void;
 }
 
 export interface IClickButtonProps {
